3-agent-tool-memory/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.3
langchain-community>=0.0.1
python-dotenv>=1.0.0
streamlit>=1.32.0
openai>=1.12.0
typing-extensions>=4.5.0
duckduckgo-search>=4.4.0
pydantic>=2.6.1
